Add Home in tabs to show all the track information
